// src/apis/models/DashboardStatsRulesCountData.ts
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class DashboardStatsRulesCountData {
  'ruleId': string;
  'hitCount': number;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'ruleId',
      baseName: 'ruleId',
      type: 'string',
      format: '',
    },
    {
      name: 'hitCount',
      baseName: 'hitCount',
      type: 'number',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return DashboardStatsRulesCountData.attributeTypeMap;
  }

  public constructor() {}
}